Modernised library for Xcode 9 / 64-bit devices.

+++ OpenID/Base64.m
@@ -41,7 +41,7 @@ static const short _base64DecodingTable[256] = {
 	char * strResult;
     
 	// Get the Raw Data length and ensure we actually have data
-	int intLength = [objData length];
+	NSUInteger intLength = [objData length];
 	if (intLength == 0) return nil;
     
 	// Setup the String-based Result placeholder and pointer within that placeholder
@@ -83,7 +83,7 @@ static const short _base64DecodingTable[256] = {
 
 + (NSData *) base64DecodeString: (NSString *) strBase64 {
 	const char * objPointer = [strBase64 cStringUsingEncoding:NSASCIIStringEncoding];
-	int intLength = strlen(objPointer);
+	unsigned long intLength = strlen(objPointer);
 	int intCurrent;
 	int i = 0, j = 0, k;
